# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "API Cafes",
    "description": "Documentacion para la API de Cafes",
    "version": "1.0.0"
  },
  "host": "localhost:8080",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/api/usuarios/registro": {
      "post": {
        "tags": [
          "Usuarios"
        ],
        "description": "Endpoint para registrar un usuario",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "correo": {
                  "example": "any"
                },
                "contraseña": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/usuarios/entrar": {
      "post": {
        "tags": [
          "Usuarios"
        ],
        "description": "Endpoint para loguear un usuario",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "correo": {
                  "example": "any"
                },
                "contraseña": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/usuarios/saldo": {
      "get": {
        "tags": [
          "Usuarios"
        ],
        "description": "Endpoint para ver el saldo de un usuario",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/transacciones/": {
      "get": {
        "tags": [
          "Transacciones"
        ],
        "description": "Endpoint para ver las transacciones de un usuario",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/transacciones/todas": {
      "get": {
        "tags": [
          "Transacciones"
        ],
        "description": "Endpoint para ver todas las transacciones",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/transacciones/recargar": {
      "post": {
        "tags": [
          "Transacciones"
        ],
        "description": "Endpoint para recargar",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "usuario": {
                  "example": "any"
                },
                "correo": {
                  "example": "any"
                },
                "importe": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/transacciones/transaccion": {
      "post": {
        "tags": [
          "Transacciones"
        ],
        "description": "Endpoint para hacer transacciones",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "usuario": {
                  "example": "any"
                },
                "correo": {
                  "example": "any"
                },
                "concepto": {
                  "example": "any"
                },
                "importe": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  }
}